restructured analysis_toolkit's child module
