+++ src/analysis_toolkit/__init__.py
@@ -1,5 +1,8 @@
-from .data_container import SamplesContainer
-from .barchart_generator import BarchartGenerator
-from .hap_network_nex_generator import HapNetNexusGenerator
-from .ml_tree_generator import MLTreeGenerator
-from .umap_generator import UmapGenerator
+# class for data storage
+from analysis_toolkit.runner_exec.data_container import SamplesData
+# class for running analysis
+from analysis_toolkit.runner_exec.runner_barchart import BarchartRunner
+from analysis_toolkit.runner_exec.runner_nexus import NexusRunner
+from analysis_toolkit.runner_exec.runner_mltree import MLTreeRunner
+from analysis_toolkit.runner_exec.runner_umap import UmapRunner
+from analysis_toolkit.runner_exec.runner_hdbscan import HdbscanRunner
