+++ src/analysis_toolkit/runner_build/base_runner.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 
-from analysis_toolkit import data_container
-from analysis_toolkit.utils import base_logger
+from analysis_toolkit.runner_build import base_logger
+from analysis_toolkit.runner_exec import data_container
 
 
 class Runner(ABC):
@@ -11,13 +11,13 @@ class Runner(ABC):
         self.sample_id_used = None
         self.parameters = {}
         self.results_dir = None
-        self.import_data(samplesdata)
+        self._import_data(samplesdata)
 
-    def import_data(self, samplesdata):
+    def _import_data(self, samplesdata):
         self.sample_data = samplesdata.sample_data
         self.sample_id_list = samplesdata.sample_id_list
 
-    def load_sample_id_list(self, sample_id_list: str = []):
+    def _load_sample_id_list(self, sample_id_list: str = []):
         if sample_id_list == []:
             base_logger.logger.info(f"No sample ID list specified. Using all {len(sample_id_list)} samples.")
             self.sample_id_used = self.sample_id_list
@@ -28,13 +28,12 @@ class Runner(ABC):
                     raise ValueError(f"Specified invalid sample ID: {sample_id}.")
             self.sample_id_used = sample_id_list
 
-    # TODO(SW): Refactor **_target() and **write_output() with the following
     @abstractmethod
-    def target(self, *args):
+    def run_write(self):
         pass
 
     @abstractmethod
-    def write_output(self, *args):
+    def run_plot(self):
         pass
 
 class SequenceRunner(ABC, Runner):
@@ -42,7 +41,13 @@ class SequenceRunner(ABC, Runner):
         super().__init__(samplesdata)
         self.units2fasta = {}
 
-    def load_units2fasta_dict(self,
+    def _filter_sequence(self, n_unit_threshold):
+        for unit, fasta in self.units2fasta.copy().items():
+            seq_num = fasta.count('>')
+            if seq_num < n_unit_threshold:
+                del self.units2fasta[unit]
+
+    def _load_units2fasta_dict(self,
             target_name: str,
             target_level: str,
             unit_level: str,
@@ -61,12 +66,16 @@ class SequenceRunner(ABC, Runner):
                     self.units2fasta[unit_name] = ""
                 self.units2fasta[unit_name] += f'>{title}\n{seq}\n'
 
+        if n_unit_threshold > 1:
+            self._filter_sequence(n_unit_threshold)
+
+
 class AbundanceRunner(ABC, Runner):
     def __init__(self, samplesdata: data_container.SamplesData):
         super().__init__(samplesdata)
-        self.units2abundance = {}
+        self.samples2abundance = {}
 
-    def load_hap_size(self, sample_id: str, hap: str) -> int:
+    def _load_hap_size(self, sample_id: str, hap: str) -> int:
         """
         Get the size of a haplotype for a given sample.
 
@@ -76,7 +85,7 @@ class AbundanceRunner(ABC, Runner):
         """
         return int(self.sample_data[sample_id].hap_size[hap])
 
-    def load_units2abundance_dict(self, sample_id: str, unit_level: str):
+    def _load_units2abundance_dict(self, sample_id: str, unit_level: str):
         """
         Get the abundance of a level for a given sample.
 
@@ -84,14 +93,15 @@ class AbundanceRunner(ABC, Runner):
         :param level: The name of the level.
         :return: A dictionary mapping level names to abundances.
         """
+        self.units2abundance = {}
         for hap, level_dict in self.sample_data[sample_id].hap2level.items():
             level_name = level_dict[unit_level]
             if level_name not in self.units2abundance:
                 self.units2abundance[level_name] = 0
-            size = self.load_hap_size(sample_id, hap)
+            size = self._load_hap_size(sample_id, hap)
             self.units2abundance[level_name] += size # e.g. {'SpA': 3, 'SpB': 4, 'SpC': 5}
     
-    def normalize_abundance(self):
+    def _normalize_abundance(self):
         """
         Normalize the abundance values in the dictionary to percentages.
 
@@ -100,3 +110,6 @@ class AbundanceRunner(ABC, Runner):
         """
         total_size = sum(self.units2abundance.values())
         self.units2abundance = {key: value/total_size * 100 for key, value in self.units2abundance.items()}
+    
+    def _update_samples2abundance_dict(self, sample_id: str):
+        self.samples2abundance[sample_id] = self.units2abundance.copy()
