+++ src/analysis_toolkit/runner_build/utils.py
@@ -2,6 +2,18 @@ import matplotlib
 import matplotlib.pyplot as plt
 import numpy as np
 
+
+def list_union(lists_to_union: list[list[str]]) -> list[str]:
+    """
+    Return the union of multiple lists as a sorted list of unique unit names.
+
+    :param lists_to_union: A list of lists to be unioned.
+    :return: A sorted list of unique elements.
+    """
+    uniq_list = list(set().union(*lists_to_union))
+    uniq_list.sort()
+    return uniq_list
+
 def get_color_hex(n: int, cmap: str = "rainbow") -> list[str]:
     color_key = plt.get_cmap(cmap)(np.linspace(0, 1, n))
     color_hex = [matplotlib.colors.to_hex(color_key[i]) for i in range(n)]
