+++ src/analysis_toolkit/runner_build/utils_sequence.py
@@ -2,7 +2,8 @@ import os
 import subprocess
 import tempfile
 
-from analysis_toolkit.utils import base_logger
+from analysis_toolkit.runner_build import base_logger
+
 
 def derep_fasta(seq_path: str, uniq_path: str, relabel: str, threads: int = 12, sizeout: bool = False) -> None:
     """
